refactor: split ConfigEditor into multiple files for better modularity
